MZ2-10 [refactor]: 기존의 next/Images를 SVG import 로 대체
기존의 next/Images 로직으로 불러오던 svg를 대체하였습니다.
